src/policy: relocate inline tests to policy_tests.rs
Same playbook. Extract 347-LOC inline test block from policy.rs to
sibling policy_tests.rs.

After: policy.rs drops from 807 → 463 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
